+++ src/app/UI/Admin/LanguageTranslateLog/LanguageTranslateLogPresenter.php
@@ -8,6 +8,7 @@ use App\Model\Admin\LanguageTranslate;
 use App\Model\Entity\LanguageTranslateEntity;
 use App\UI\Accessory\Admin\PresenterTrait\RequireLoggedUserTrait;
 use App\UI\Accessory\Admin\PresenterTrait\StandardTemplateTrait;
+use App\UI\Accessory\Admin\Submenu\SubmenuFactory;
 use App\UI\Admin\LanguageTranslateLog\DataGrid\LanguageTranslateLogDataGridEntityFactory;
 use Nette\Application\UI\Presenter;
 use Nette\Utils\Json;
@@ -21,8 +22,21 @@ class LanguageTranslateLogPresenter extends Presenter
         private readonly LanguageTranslate $languageTranslateModel,
         private readonly LanguageTranslateLogDataGridEntityFactory $dataGridEntityFactory,
         private readonly DataGridFactory $dataGridFactory,
+        private readonly SubmenuFactory $submenuFactory,
     ) {
         parent::__construct();
+    }
+
+    public function actionDefault(): void
+    {
+        $this->submenuFactory->addMenu($this->translator->translate('menu_languageTranslateLog_markAllFinished'), 'markAllFinished');
+    }
+
+    public function actionMarkAllFinished(): void
+    {
+        $count = $this->languageTranslateModel->markAllUnfinished(new \DateTime());
+        $this->flashMessage($this->translator->translate('flash_languageTranslateLog_marked', ['count' => $count]));
+        $this->redirect('default');
     }
 
     public function actionDetail(int $id): void